fix: define virtual methods

+++ navigation_filter_library/include/sensor_model.hpp
@@ -11,7 +11,7 @@ private:
     //Twist parameters
     float update_time_;
     Eigen::Matrix4f transform_;
-    Eigen::MatrixXf covariance_;
+    Eigen::MatrixXd covariance_;
 
 public:
     SensorModel(Eigen::Matrix4f transform, Eigen::VectorXd noise, float dt);
@@ -20,7 +20,7 @@ public:
 
     float get_update_time();
     Eigen::Matrix4f get_transform();
-    Eigen::MatrixXf get_covariance();
+    Eigen::MatrixXd get_covariance();
 
     virtual Eigen::MatrixXf get_jacobian(ErrorState state);
     virtual Eigen::MatrixXf get_noise_jacobian(ErrorState state);
